requirements.txt
flask>=3.0.0
twilio>=9.0.0
openai>=1.30.0
python-dotenv>=1.0.0
googlesearch-python>=1.2.3
beautifulsoup4>=4.12.0
requests>=2.31.0
gunicorn>=21.2.0
redis>=5.0.0
lxml>=4.9.0
